requirements.txt
pandas
sqlalchemy
mysql-connector-python
google-cloud-bigquery
python-dotenv
pytest
pyarrow
pandas-gbq
